perf(bundle): extract inline ModulePanel panels into next/dynamic chunks
Backlog: load-time push. PR #300 lazy-loaded the four sub-panels that
already lived in their own files. The remaining inline definitions
(TarskiPanel, ParetoPanel, CopilotInterdictionResults,
SnapshotIndicator + their co-located helpers — AxiomIcon,
ProofTraceList, CritSparklineChart, CritSparkline, shortenEventLabel,
CriticalityCard, type CriticalityEmptyState) couldn't be deferred
without first extracting them. ~2.5K LOC + heavy estimator-lib
transitive deps shipped on every initial paint, even though Spirtes
is the default tab and needs none of them.

Three new files under src/components/modules/:
- CopilotInterdictionResults.tsx (~253 LOC) — Pearl tab
- TarskiPanel.tsx (~599 LOC) — Tarski tab; brings AXIOM_LIBRARY +
  scoreAxiomRelevance with it
- ParetoPanel.tsx (~1.78K LOC) — Pareto tab, the heaviest of the
  three. Co-locates SnapshotIndicator (named export so outer
  ModulePanel dynamic-imports both from the same chunk), sparkline
  components, shortenEventLabel, CriticalityCard, the
  CriticalityEmptyState type. Pulls the estimator-lib transitive
  deps with it: lppls-fit, ph-fit, pareto-relevance-bootstrap,
  pareto-relevance-reference, moran, t1d-estimator-inputs.

ModulePanel.tsx: removed all four inline definitions, trimmed 14
now-unused lib-level imports, added four `next/dynamic`
declarations (the SnapshotIndicator one uses
`.then(m => ({ default: m.SnapshotIndicator }))` so it shares the
ParetoPanel chunk).

ModulePanel.tsx shrank 3384 LOC → 821 LOC. Default Spirtes-tab
first paint is now CascadeHeader + TrinityPanel + DiscoveryRunsPanel
(+ optional TissueCohortView) — all heavy regime-gate / criticality-
card code deferred to first-tab-visit.

Drive-by: fixed two missing fields on the worker test fixture
(`isConfounded`, `inconsistentEdges`, `restrictedNodes`) that tsc
started flagging since PR #404 landed.

Verification: tsc clean (same pre-existing inherited errors); lint
has 2 errors that are the SAME pre-existing set-state-in-effect +
rules-of-hooks issues from the original inline definitions (just
moved to their new files, identical code); vitest 1321 / 1321 pass.
